Updated Mod_Contact_User.tcKDTest to post a color coded warning to the log prior to the intrinsic failure generated by TC for this missing or misconfigured driver.
The needed data source driver is not installed or is not configured properly.
To read data, install the 64-bit version of Microsoft Office 12.0 Access Database Engine OLE DB Provider. It is part of Microsoft Access Database Engine 2010 Redistributable. You can find it on the Microsoft website. Alternatively, you can try using the 32-bit version of TestExecute.